requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
openai==0.28.1
numpy==1.26.4
gunicorn==22.0.0
yt-dlp==2024.07.09
eventlet==0.36.1
requests==2.31.0